tests/test_kintone_api.py
import pytest
from unittest.mock import patch
from unittest.mock import mock_open
from kintone.kintone_api import KintoneApi

def test_constructor():
    api = KintoneApi(domain='example.kintone.com', user='username', password='password')
    assert api.url_prefix == 'https://example.kintone.com'

@patch('urllib3.PoolManager.request')
def test_get_method(mock_request):
    mock_request.return_value.status = 200
    mock_request.return_value.data = b'{"message": "success"}'
    api = KintoneApi(domain='example.kintone.com', user='username')
    result = api.get('/test')
    assert result['message'] == 'success'

@patch('urllib3.PoolManager.request')
def test_post_method(mock_request):
    mock_request.return_value.status = 200
    mock_request.return_value.data = b'{"message": "success"}'
    api = KintoneApi(domain='example.kintone.com', user='username')
    result = api.post('/test', {'key': 'value'})
    assert result['message'] == 'success'

@patch('urllib3.PoolManager.request')
def test_delete_method(mock_request):
    mock_request.return_value.status = 200
    mock_request.return_value.data = b'{"message": "success"}'
    api = KintoneApi(domain='example.kintone.com', user='username')
    result = api.delete('/test')
    assert result['message'] == 'success'

@patch('urllib3.PoolManager.request')
@patch('builtins.open', new_callable=mock_open, read_data=b'file_content')
def test_post_file_method(mock_file, mock_request):
    mock_request.return_value.status = 200
    mock_request.return_value.data = b'{"message": "success"}'
    api = KintoneApi(domain='example.kintone.com', user='username')
    result = api.post_file('/test', file_path='path/to/file.txt')
    assert result['message'] == 'success'